stock-price-fetcher/pyproject.toml
[project]
name = "stock-price-fetcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.2",
    "yfinance>=1.2.0",
]
